handlers, responce helpers

+++ internal/model/stat.go
@@ -1,6 +1,6 @@
 package model
 
 type Stat struct{
-	MaxPress int `json:"max-press"`
-	PersentBetter uint16 `json:"persent-better"`
+	MaxPress float32`json:"max-press"`
+	PersentBetter float32 `json:"persent-better"`
 }
